refactor: requestLine 객체 추가
