+++ app/src/main/java/nextstep/jwp/controller/Controller.java
@@ -1,7 +1,7 @@
 package nextstep.jwp.controller;
 
-import nextstep.jwp.http.HttpRequest;
-import nextstep.jwp.http.HttpResponse;
+import nextstep.jwp.http.request.HttpRequest;
+import nextstep.jwp.http.response.HttpResponse;
 import nextstep.jwp.exception.ResourceNotFoundException;
 
 import java.io.File;
