+++ app/src/main/java/nextstep/jwp/controller/RegisterController.java
@@ -1,6 +1,8 @@
 package nextstep.jwp.controller;
 
 import nextstep.jwp.http.*;
+import nextstep.jwp.http.request.HttpRequest;
+import nextstep.jwp.http.response.HttpResponse;
 import nextstep.jwp.service.UserService;
 
 public class RegisterController extends Controller {
@@ -12,9 +14,9 @@ public class RegisterController extends Controller {
 
     @Override
     public boolean canHandle(final HttpRequest httpRequest) {
-        final String httpMethod = httpRequest.getHttpMethod();
+        final HttpMethod httpMethod = httpRequest.getHttpMethod();
         final String path = httpRequest.getPath();
-        return HttpMethod.isPost(httpMethod) && "/register".equals(path);
+        return httpMethod.isPost() && "/register".equals(path);
     }
 
     @Override
