+++ app/src/main/java/nextstep/jwp/http/RequestHandler.java
@@ -3,6 +3,8 @@ package nextstep.jwp.http;
 import nextstep.jwp.controller.*;
 import nextstep.jwp.exception.NoMatchingControllerException;
 import nextstep.jwp.exception.ResourceNotFoundException;
+import nextstep.jwp.http.request.HttpRequest;
+import nextstep.jwp.http.response.HttpResponse;
 import nextstep.jwp.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
