+++ app/src/test/java/nextstep/jwp/controller/RegisterControllerTest.java
@@ -1,6 +1,10 @@
 package nextstep.jwp.controller;
 
 import nextstep.jwp.http.*;
+import nextstep.jwp.http.request.HttpRequest;
+import nextstep.jwp.http.request.HttpRequestBody;
+import nextstep.jwp.http.request.HttpRequestHeader;
+import nextstep.jwp.http.response.HttpResponse;
 import nextstep.jwp.service.UserService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
