+++ app/src/test/java/nextstep/jwp/controller/StaticResourceControllerTest.java
@@ -1,6 +1,9 @@
 package nextstep.jwp.controller;
 
 import nextstep.jwp.http.*;
+import nextstep.jwp.http.request.HttpRequest;
+import nextstep.jwp.http.request.HttpRequestHeader;
+import nextstep.jwp.http.response.HttpResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
