+++ app/src/test/java/nextstep/jwp/http/RequestHandlerTest.java
@@ -1,6 +1,7 @@
 package nextstep.jwp.http;
 
 import nextstep.jwp.exception.NoMatchingControllerException;
+import nextstep.jwp.http.response.HttpResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
